reverted test

+++ EFCore.BulkExtensions.Tests/EFCoreBatchTest.cs
@@ -52,14 +52,13 @@ namespace EFCore.BulkExtensions.Tests
                 context.Items.BatchDelete();
                 context.BulkDelete(context.Items.ToList());
 
-                switch (databaseType)
+                if (databaseType == DbServer.SqlServer)
                 {
-                    case DbServer.SqlServer:
-                        context.Database.ExecuteSqlRaw("DBCC CHECKIDENT('[dbo].[Item]', RESEED, 0);");
-                        break;
-                    case DbServer.Sqlite:
-                        context.Database.ExecuteSqlRaw("DELETE FROM sqlite_sequence WHERE name = 'Item';");
-                        break;
+                    context.Database.ExecuteSqlRaw("DBCC CHECKIDENT('[dbo].[Item]', RESEED, 0);");
+                }
+                if (databaseType == DbServer.Sqlite)
+                {
+                    context.Database.ExecuteSqlRaw("DELETE FROM sqlite_sequence WHERE name = 'Item';");
                 }
             }
         }
